Performance test (#279)
* Delete die code

* Add transaction test

* Move standard_transaction_test to test dir

* Add performance test for NewBlockTpl function
