--- cmd/bytomcli/test/transaction_test
@@ -1,106 +1,0 @@
-#!/usr/bin/env bash
-
-set -eo pipefail
-
-help() {
-  cat << EOF
-usage: $0 [OPTIONS]
-
-    -s --sender_account_id      The account ID of transaction sender
-    -A --asset_id               The asset ID
-    -r --receiver_program       The receiver program
-    -t --type                   Type of test: 'issue' or 'spend', the default is 'issue'
-    -a --amount                 The amount of asset
-    -p --password               The password of sender_account_id
-EOF
-}
-
-if [ $# -eq 0 ]; then
-  help
-  exit 1
-fi
-
-POSITIONAL=()
-while [[ $# -gt 0 ]]
-do
-key="$1"
-
-case $key in
-    -s|--sender_account_id)
-    sender_account_id="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    -r|--receiver_program)
-    receiver_program="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    -p|--password)
-    password="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    -A|--asset_id)
-    asset_id="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    -t|--type)
-    test_type="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    -a|--amount)
-    amount="$2"
-    shift # past argument
-    shift # past value
-    ;;
-    *)
-    echo "Unknown flag: $2"
-    exit 1
-    ;;
-    *)    # unknown option
-    POSITIONAL+=("$1") # save it in an array for later
-    shift # past argument
-    ;;
-esac
-done
-set -- "${POSITIONAL[@]}" # restore positional parameters
-
-get_data() {
-cat << EOF
-$1
-EOF
-}
-
-export PATH="$GOPATH/src/github.com/bytom/cmd/bytomcli:$PATH"
-
-issue_test() {
-  echo "Issue test:"
-  sign_tmpl=$(bytomcli build-transaction $sender_account_id $asset_id $amount -t issue | grep "raw_transaction" | tr -d '\n')
-  sub_tmpl=$(bytomcli sign-transaction "$(get_data "$sign_tmpl")" -p $password | grep "transaction" | tr -d '\n')
-  bytomcli submit-transaction "$(get_data "$sub_tmpl")" | tr -d '\n'
-  echo
-  [ $? -eq 0 ] && echo "Issue test passed" || echo "Issue test failed"
-}
-
-spend_test() {
-  echo "Spend test:"
-  sign_tmpl=$(bytomcli build-transaction "$sender_account_id" "$asset_id" "$amount" -t spend -r "$receiver_program" | grep "raw_transaction" | tr -d '\n')
-
-  sub_tmpl=$(bytomcli sign-transaction "$(get_data "$sign_tmpl")" -p $password | grep "transaction" | tr -d '\n')
-
-  bytomcli submit-transaction "$(get_data "$sub_tmpl")" | tr -d '\n'
-  echo
-  [ $? -eq 0 ] && echo "Spend test passed" || echo "Spend test failed"
-}
-
-if [[ $test_type == 'issue' ]]; then
-  issue_test
-elif [[ $test_type == 'spend' ]]; then
-  spend_test
-else
-  echo "Invalid test type"
-  exit 1
-fi
