--- cmd/bytomd/run_test.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"os"
-	"testing"
-)
-
-func TestMain(m *testing.M) {
-	os.Exit(m.Run())
-}
